requirements.txt
python-dateutil
requests
prettytable

# for timeline
plotly
numpy

# for colors
seaborn

# lexical analysis
Distance